fix: Remove magic field type mismatch and unused audio handlers
- Change frame_t.magic and multi_source_frame_t.magic from uint32_t to uint64_t to match MAGIC_FRAME_VALID constant (0xA5C11C4A1ULL)
- Remove unused client audio handlers: handle_client_audio_opus, handle_client_audio
- Remove unused server audio handlers: handle_server_audio_opus, handle_server_audio
- These handlers were not in the dispatch tables and only caused compiler warnings

This eliminates [-Wtautological-constant-out-of-range-compare], [-Wconstant-conversion], and [-Wunused-function] warnings.

+++ lib/ringbuffer.h
@@ -386,7 +386,7 @@ void ringbuffer_clear(ringbuffer_t *rb);
  */
 typedef struct {
   /** @brief Magic number to detect corruption (FRAME_MAGIC when valid) */
-  uint32_t magic;
+  uint64_t magic;
   /** @brief Actual size of frame data in bytes */
   size_t size;
   /** @brief Pointer to frame data (not owned by this struct) */
@@ -403,7 +403,7 @@ typedef struct {
  */
 typedef struct {
   /** @brief Magic number to detect corruption (FRAME_MAGIC when valid) */
-  uint32_t magic;
+  uint64_t magic;
   /** @brief Client ID that sent this frame */
   uint32_t source_client_id;
   /** @brief Frame sequence number for ordering */
